Abstract annual reports into data file
